# perf: use buf in Args instead of bytebufferpool